[UAR-262] Update version of trust details page
