+++ src/config/index.ts
@@ -229,7 +229,7 @@ export const UPDATE_REVIEW_BENEFICIAL_OWNER_OTHER_URL = UPDATE_AN_OVERSEAS_ENTIT
 export const UPDATE_CONFIRM_TO_REMOVE_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_CONFIRM_TO_REMOVE_PAGE;
 export const UPDATE_BENEFICIAL_OWNER_TYPE_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_BENEFICIAL_OWNER_TYPE_PAGE;
 export const UPDATE_TRUSTS_SUBMISSION_INTERRUPT_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_TRUSTS_SUBMISSION_INTERRUPT_PAGE;
-export const UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_TRUSTS_TELL_US_ABOUT_IT_PAGE + OPTIONAL_ID_PARAM;
+export const UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_TRUSTS_TELL_US_ABOUT_IT_PAGE;
 export const UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_PAGE;
 export const UPDATE_TRUSTS_ASSOCIATED_WITH_THE_OVERSEAS_ENTITY_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_TRUSTS_ASSOCIATED_WITH_THE_OVERSEAS_ENTITY_PAGE;
 export const UPDATE_CHECK_YOUR_ANSWERS_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_CHECK_YOUR_ANSWERS_PAGE;
