+++ test/__mocks__/text.mock.ts
@@ -234,3 +234,5 @@ export const NO_CHANGE_REVIEW_STATEMENT_BENEFICIAL_OWNER_STATEMENTS_CEASED_TITLE
 export const BENEFICIAL_OWNER_HEADING = "Beneficial owners";
 export const CONFIRMATION_AGENT_SPECIFIC_TEXT = "Once we have received the statement from the agent";
 export const CONFIRMATION_UPDATE_TEXT = "if the update statement has been accepted or rejected";
+export const UPDATE_TELL_US_ABOUT_TRUST_HEADING = "Tell us about the trust";
+export const UPDATE_TELL_US_ABOUT_TRUST_QUESTION = "What is the name of the trust?";
