+++ test/controllers/update/update.trusts.tell.us.about.it.controller.spec.ts
@@ -4,6 +4,7 @@ jest.mock('../../../src/utils/application.data');
 jest.mock('../../../src/middleware/authentication.middleware');
 jest.mock('../../../src/middleware/company.authentication.middleware');
 jest.mock('../../../src/middleware/service.availability.middleware');
+jest.mock('../../../src/utils/save.and.continue');
 
 import { beforeEach, jest, test, describe } from '@jest/globals';
 import request from 'supertest';
@@ -22,8 +23,9 @@ import { getApplicationData } from '../../../src/utils/application.data';
 import { isActiveFeature } from '../../../src/utils/feature.flag';
 
 import { APPLICATION_DATA_MOCK } from '../../__mocks__/session.mock';
-import { PAGE_TITLE_ERROR, PAGE_NOT_FOUND_TEXT } from '../../__mocks__/text.mock';
+import { PAGE_TITLE_ERROR, PAGE_NOT_FOUND_TEXT, UPDATE_TELL_US_ABOUT_TRUST_HEADING, UPDATE_TELL_US_ABOUT_TRUST_QUESTION, ERROR_LIST } from '../../__mocks__/text.mock';
 import { saveAndContinueButtonText } from '../../__mocks__/save.and.continue.mock';
+import { saveAndContinue } from '../../../src/utils/save.and.continue';
 
 const mockGetApplicationData = getApplicationData as jest.Mock;
 mockGetApplicationData.mockReturnValue( APPLICATION_DATA_MOCK );
@@ -40,6 +42,8 @@ mockServiceAvailabilityMiddleware.mockImplementation((req: Request, res: Respons
 const mockIsActiveFeature = isActiveFeature as jest.Mock;
 mockIsActiveFeature.mockReturnValue(true);
 
+const mockSaveAndContinue = saveAndContinue as jest.Mock;
+
 describe('Update - Trusts - Tell us about the trust', () => {
   beforeEach(() => {
     jest.clearAllMocks();
@@ -47,19 +51,20 @@ describe('Update - Trusts - Tell us about the trust', () => {
 
   describe('GET tests', () => {
     test('when feature flag is on, page is returned', async () => {
-      mockIsActiveFeature.mockReturnValue(true);
+      mockIsActiveFeature.mockReturnValueOnce(true);
 
       const resp = await request(app).get(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL);
 
       expect(resp.status).toEqual(200);
-      expect(resp.text).toContain('Tell us about the trust');
+      expect(resp.text).toContain(UPDATE_TELL_US_ABOUT_TRUST_HEADING);
+      expect(resp.text).toContain(UPDATE_TELL_US_ABOUT_TRUST_QUESTION);
       expect(resp.text).toContain(UPDATE_TRUSTS_SUBMISSION_INTERRUPT_URL);
       expect(resp.text).toContain(saveAndContinueButtonText);
       expect(resp.text).not.toContain(PAGE_TITLE_ERROR);
     });
 
     test('when feature flag is off, 404 is returned', async () => {
-      mockIsActiveFeature.mockReturnValue(false);
+      mockIsActiveFeature.mockReturnValueOnce(false);
 
       const resp = await request(app).get(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL);
 
@@ -69,17 +74,37 @@ describe('Update - Trusts - Tell us about the trust', () => {
   });
 
   describe('POST tests', () => {
-    test('when feature flag is on, redirect to individuals or entities involved in trust page', async () => {
-      mockIsActiveFeature.mockReturnValue(true);
+    test('when feature flag is on and no data posted, re-render page with validation error', async () => {
+      mockIsActiveFeature.mockReturnValueOnce(true);
 
-      const resp = await request(app).post(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL);
+      const resp = await request(app).post(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL).send({});
 
+      expect(resp.status).toEqual(200);
+      expect(resp.text).toContain(ERROR_LIST);
+      expect(resp.text).toContain(UPDATE_TELL_US_ABOUT_TRUST_HEADING);
+      expect(resp.text).toContain(UPDATE_TELL_US_ABOUT_TRUST_QUESTION);
+    });
+
+    test('when feature flag is on and posting valid data, redirect to update-trusts-individuals-or-entities-involved page', async () => {
+      mockIsActiveFeature.mockReturnValueOnce(true);
+
+      const resp = await request(app).post(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL).send({
+        name: 'Trust name',
+        createdDateDay: '08',
+        createdDateMonth: '07',
+        createdDateYear: '2023',
+        beneficialOwnersIds: '45e4283c-6b05-42da-ac9d-1f7bf9fe9c85',
+        hasAllInfo: '0',
+        trustId: ''
+      });
+
+      expect(mockSaveAndContinue).toHaveBeenCalledTimes(1);
       expect(resp.status).toEqual(302);
       expect(resp.header.location).toEqual(UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL);
     });
 
     test('when feature flag is off, 404 is returned', async () => {
-      mockIsActiveFeature.mockReturnValue(false);
+      mockIsActiveFeature.mockReturnValueOnce(false);
 
       const resp = await request(app).post(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL);
